+++ src/main/java/roomescape/domain/reservation/ReservationWithRank.java
@@ -1,5 +1,6 @@
 package roomescape.domain.reservation;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalTime;
 
@@ -10,6 +11,6 @@ public record ReservationWithRank(
         LocalTime time,
         Status status,
         String paymentKey,
-        Integer amount,
+        BigDecimal amount,
         long waitingRank) {
 }
